Summary: Add kluge to handle coroutine case of actormap-churn-run!
The problem is that if someone puts an await during the submitted thunk
it would return one value instead of the list of values.

This defines a type that will be recognized as being the multival
return kluge, and in that case uses applies to values, otherwise just
returns the value as-is.