Add initial .gitlab-ci.yml file
This adds a CI configuration file. This defines a "test" stage which
runs the bootstrap, configure, make and then make check. This currently
takes a long time because the docker image being used is on Guix 1.3.0
which does not have the required "guix shell" command, to fix that we're
doing a "guix pull".

Fixes #1