Add some information to the README and change to define-values
This fixes the coverage command to use the `--debug` command with
guile to generate the coverage. Without this command, no coverage data
is actually generated. It also includes better information regarding
generating the HTML report with `genhtml`.

Further, this changes the usage of `call-with-values` to
`define-values` within the `run-coverage` function and adds some
documentation to each function.