Comment out the record type printer for actormaps, at least for now
This is a loss of quality of life, but currently we're experiencing
some sort of bug.  I'm not sure if it's expected, but somehow when
printing out transactormaps at the REPL, occasionally we're getting
exceptions like:

```
scheme@(guile-user) [1]> ,bt
In ice-9/boot-9.scm:
  1747:15  6 (with-exception-handler #<procedure 7f472d08b240 at ice-9/boot-9.scm:1831:…> …)
   260:13  5 (for-each #<procedure 7f472d2e94a0 at system/repl/repl.scm:207:34 (v)> _)
In system/repl/common.scm:
   259:11  4 (repl-print _ _)
In unknown file:
           3 While executing meta-command:
Wrong type to apply: #<syntax-transformer actormap-metatype-name>
```

I'll admit that I find this to be surprising and confusing both.
I have no idea why we'd be hitting this kind of exception, but ok
I guess...

The main downside here is that we now get more verbose representations
of actormaps at the REPL.  Given that most users are operating on
vats and not touching actormaps directly, this isn't so bad, but it's
not what I'd like long run.  It beats getting an error when trying to
debug an error, though.