Fix vats launched from geiser buffer evaluations freezing
The purpose of this is to prevent issues where a user hacking with
Geiser's buffer evaluation commands (eg C-x C-e) launches a vat, and
things weirdly break... because geiser sets redirects output so that
it can capture it to display to the user when hacking that way, but
those ports are closed at the end of the evaulation.  But since the
vat would run in its own fiber/thread, any attempts to write to
output/error ports would throw an exception.  This redirects them
"back".

 - Add base-io-ports.scm to "keep track" of what the original ports
   were.
 - Add base-io-ports.scm to hall.scm (and default-vat-scheduler.scm
   too, which should have been there already.)
 - Add keyword arguments to `spawn-vat-proc` and `spawn-vat-fiber`
   which allow for controlling the "dyamic launching" of a vat.
   This could thus either be overruled or... something else
   interesting could be done.
 - Add a docstring for spawn-vat-fiber and some comments explaining
   all of this.