Correct test for promise pipelining
Previous version of test was using `#:promise? #t` alongside a `#:catch`,
but `#:catch` will instead return its value, which in this case was
actually undefined/void because of the `set!`.

This revised version also allows for ensuring that #:catch clause of
promise pipelining is run in appropriate case too (and uhoh, it
isn't, so good thing we're checking!)