Introduce a function based approach to marshalling records for syrup
This removes the old macro for defining both the record
and (un)marshalling functions and replaces it with a function which
when given a specific record type and name, generates the
corrisponding (un)marshalling functions for the record.

This converts all uses of define-recordable to the newer approach and
removes the define-recordable macro.