Fix a bug with encoding syrup objects
There was a bug where encoding pairs, syrup records and sets could
result in a syrup encoding error when using marshallers. The
`syrup-encode` function is the outward interface that the user uses and
passes in the set of marshallers for the data they wish to encode. There
is an inner `encode` function which has all the marshallers passed in
available in the current scope.

The recursive encoding needs to use the marshallers passed in by the
user to encode the data. This requires the syrup encoding library to use
the `encode` which has those available, opposed to the outward
`syrup-encode` which would try to encode the inner data without any
marshallers in scope.