Convert flakey fix to use fibers channels
When testing message sending between vats, we need to reliably get the
result of the message. Before we were just sleeping which caused an
issue as it hadn't always resolved the promise. Then we moved to a
longer timeout with polling periodically to try and shortcircuit the
timeout making the tests fast.

This solution however uses fibers channels to solve the issue. It'll
setup a fibers channel which blocks waiting for an answer from the vat
code, the vat code will send a message when the promise resolves
ensuring we wait as long as needed, but no longer than we require.

This should make the vat tests much less flakey and a similar method can
be used for other tests in the future.