Rewrite and remove some of structs-urls
This removes for now support for `ocapn-cert` and `ocapn-bearer-union`
URIs. These two were not properly supported in the code base and are
not used anywhere.

It changes the URIs to use ocapn:// vs. ocapn:. This makes more sense
for the scheme to work like this, as well as improves URI parsability
with the language libraries.

This also re-writes converting from URI structs to strings and back
using guile's URI parsing libraries. Fixes #29