Rename structs-urls to ids & various fixes
This does several big things, including renaming structs-urls to ids
to reflect better what they are.

This also backports a fix to goblins which causes the registered name
in URIs to fail if the "tld" portion of the name begins with an
integer. This doesn't follow the RFC (RFC3986), and also causes
problems with onion addresses in OCapN IDs.

Finally, this adds two additional tests which were missing
previously. These are:

- Testing from an ocapn-machine->string
- Testing from an ocapn-sturdyref->string