Improve crypto signatures and verification in captp
Handoffs currently don't work, they pass through the
<desc:sig-envelope> with the handoff-give inside, this is because we
never implemented proper crypto and had stub 'TODOs in our
structs. This caused the `signed-handoff-give?` function to report
that it wasn't a signed-handoff-give.

This provides several helper functions in the crypto utils to handle
signing and verifying. It begins to use these functions within the
captp code.