Protect ocapn ids from leakage by providing printer
Ocapn IDs provide a capability to access and use a machine or object on
a specific machine. The information contained in them should not be
leaked in stack traces or elsewhere as it could give away this
capability unintentially. This protects them by redacting the
information needed to use them.

Fix #56