Allow for installing refrs are specific swiss-nums
If you want an actor to live at a specific swiss-num (if for example you
want it to remain at a specific sturdyref), you need the ability to
register a refr at a given swiss-num.

This now allows you to optionally provide a swiss-num to register a refr
at, otherwise it will generate one for you.