Rename ocapn-machine to node and machine's address to designator
This renames ocapn-machine to ocapn-node to better match the draft OCapN
Locators draft specification. The rename from machine to node was made
as there can be several OCapN nodes/machines on a single
computer/machine, since machine is a common place word, overloading this
with a specific OCapN meaning was less than ideal. The word node was
chosen as it's a node (machine) within a graph (network).

ocapn-node-address was renamed to ocapn-node-designator to again be more
in line with the OCapN Locators draft specification. The
address/designator is often not an address in the conventional sense but
a self-authenticating designator so designator is a better label for this.