Relay netlayer with object flow
This is a relay netlayer (ie yo dawg I heard you like ocapn so I put
an ocapn in your ocapn and now you can message while you message).
This allows for using objects to send messages, allowing for relaying
messages and allowing a client-server model to have the illusion of a
peer to peer model.  Federation is built-in and easy thanks to us
using captp already under the hood; users do not have to be on the
same relay node.

Does not yet support end-to-end encryption, but does support the
object-based relay behavior.