Run setup command on netlayer after it's installed
This step was probably omitted because we hadn't really had a reason
to use the install-netlayer command since netlayers moved into a
multi-step setup process when their design was being finalized.
Anyway, all the other netlayers get set up one by one when the mycapn
is being set up, so this is necessary for this one as well... without
it, the netlayer doesn't know how to speak to the
connection-establisher.