Remove updating guix from the CI
Back when the CI was added, we used `guix shell -m manifest.scm` which
was not part of a released Guix version. This is no longer the case and
we shouldn't need to pull the latest Guix changes to run the CI which
can cause the test runner to timeout.