Fix relay-utils and its tests
Unfortunately we merged two PRs that both passed CI but interacted in
an incompatible way.  We now expect the "enliven" capability to be a
facet of a goblins mycapn object rather than an enclosed procedure.
This updates the relay-utils and test-relay-utils code to match that
expectation.

In the process it also updates the expectation that the register
capability should be a faceted object too.  Unfortunately this also
results in some ugliness because the facet needs to enclose
information about the netlayer name; whether or not that's
aesthetically the nicest thing to do is something we can discuss after
this is merged and tests are passing again.