Provide function to partial persist objects in actormap
Previously we just had the ability to take an actormap and serialize all
objects within it from a given set of roots, this was done with the
actormake-take-portrait function. This still exists, but in addition to
that there is a new function which returns a lambda that will produce a
portrait of just a single object within an already depicted graph.