Improve the actor upgrade functionality
This improves the way actors can be upgraded by having it be possible to
have the constructor be a boxed constructor. When an upgrade wants to
happen, then you can simply set the inner value of the box and the
constructor "refrs" (or boxes) remain eq? to each other.

This commit assumes all actors defined with the define-actor macro want
to have this redefinable quality, maybe this should be changed.