Reduce promises in resturation by using local-links
When restoring an object graph we create a promise pair for each
object within the graph and when an object is restored with another
objects refr, we simply hand it the vow to the object as it might not
have been restored yet.

This is especially important when handling graphs with cycles or where
objects refer to themselves as it is impossible to ensure they are
restored prior to restoring the objects which hold references to them.

The way around this is to continue to do the above with the promise
pairs but hand restored objects a refr to a mactor:local-link (actor
symlink). Initially it points to the vow so messages sent to it upon
construction are handled properly but once all the graph is restored
and we fulfill each vow, we also swap the local-link to point to the
actual object, rather than the vow.

This still means uses of the refr on construction must be written in
an asynchronous way, but it allows you to use the refr like any other
in the invoked behavior.