Fix bug in restoring actors around message dispatch
Before during resturation, for each actor it rehydrated it in its own
actormap-run!. When this happened the messages it sent were not being
dispatched and so promises made then never resolved.

When restoring the actors it now does it entirely in one transaction
instead of piece by piece and it does it in a churn so by the end of
it all the messages should be handled.