Freeze actor-lib
When an actor is frozen it's defined without the "redefinable-object"
wrapper. The lookup of the inner actor constructor and extra space is
negligable, but we can get rid of it entirely by freezing them.

Since these are provided as part of actor-lib and not likely to be
hacked upon by the user, freeze them.