Add upgrade path for aurie vat roots
There are times where you may wish to upgrade the object graph's roots
themselves, rather than the individual actors. Previously that hasn't
been possible, this introduces an upgrade path similar to that which
is provided for actors.

The #:version and #:upgrader keywords are now accepted on
spawn-persistent-vat allowing you to upgrade roots.

Fix #221