Unenclose the syscaller
This moves the syscaller out of one giant closure with an unnecessary
amount of indirection to an actual datastructure with procedures which
operate upon it.

This should reduce a certain amount of dispatch work, but most
importantly, opens the door for better optimizations and improves
readability of core.scm overall.