Fix junit XML file
The junit XML file was outputting broken XML (actually pretty much
nothing in addition to it being broken). This PR fixes it.

We were expecting combine-junit-xml-files (a procedure) to return the
sxml, not write it out to the file. Because of this we were
passing *undefined* to sxml->xml overriting what combine-junit-xml had
written. Fixes by returning what we expected from the procedure.