Fix thread-safety issue for CapTP GC
Prior to this, both the OCapN vat and the GC loop were trying to modify
the spare-imports-count hashmap. This could cause a potential problem
where both tried to modify it at the same time. This fixes that by only
allowing the GC loop to modify it.

This also makes the GC operations verbs to align more with the increment
verb.