Reconnect to peers with previous severed connection
Before this, if you had a connection to a peer and there had been a
severence, an enliven wouldn't create a new connection to that peer,
instead you'd recieve a severence error. This fixes that by having
connections clean themselves up from the sessions hashes when a
severence occurs.

Fix #657