Fix crossed hellos
The crossed hellos implementation often did not detect when it was in
the crossed hellos situation, as the mitigation object which is there to
help detect and mitigate it was removed too early. This fixes that by
now only removing them in two places:

1. If we are an inbound connection and we're aborting, we remove the
mitigator because it's done its job and isn't needed anymore.

2. At the end of a session when its severed.