Make nested with-vat non-blocking
with-vat is generally only for bootstrapping, nesting them isn't
desirable. However new users do this and it unexpectedly blocks their
vat as `with-vat`/`call-with-vat` does a blocking fibers `get-message`.
This now checks if we have a syscaller and if so returns a promise and
dispatches the thunk to a new fiber.

Fix #664