Fix reconnect prelay test after new with-vat behavior
Before we were nesting with-vat, not a great practice to begin with. Now
we have introduced the behavior of returning promises in that situation,
the code doesn't work as it's expecting multiple values returned. This
fixes that by moving the call which does the with-vats outside of the
testing vat.