Fix serializing settled near promise refrs
Before when we encountered a settled promise refr, we'd check if it is a
local refr and if so add it to our set of objects within the portrait
graph. Unfortunately this meant local promises which contained far
local-refrs were being added to the portrait graph on other vats.

This change lets the process-one procedure do all the heavy lifting for
all promise values and passes through both near and far refrs, encasing
all other values. Because process-one was already doing the right thing
distinguishing near and far refrs this also fixes the issue faced.