Fix aurie not persisting objects after rehydration
Aurie vats use a hashmap val->slot-ref (provided by aurie) to see if a
changed refr appears in the persistence graph, and therefore needs
persisting. When waking a vat backup, it initially gets the roots, but
never asks aurie to build up this val->slot-ref hashmap. This means that
unless the object is a root, the vat will not consider objects part of
the portrait graph and not persist changes to them.

This fixes it so that we still do vat-take-portrait for newly spawned
vats (to get a full portrait to disk), but for newely restored vats, we
just call out to aurie asking it to build this val->slot-refr hashmap up
for us to use later.