Fix restoring multiple far refrs to same object
When aurie encounters far refrs it needs to restore, it builds up a
hashmap with resolvers to restore a given far-refr. Unfortunately, this
did not check to see if it's already encountered the same refr before
and would clobber the resolver which was in the table each time it was
referenced causing promises to far-refrs which would never restore.

This fixes it by building up a new table with the vows used far-refrs
and giving out the same vow for each far refr, not clobbering the
resolver.