Add better support for OCapN test suite
The OCapN test suite could before be run using the onion netlayer,
but isn't ideal to be used in CI or even outside of CI contexts. This
adds the really basic tcp-testing-only netlayer for OCapN testing
purposes only. It also adds a CI job to run this.

This in addition makes a breaking change to how hints are encoded,
using a ghash (transmitted as an OCapN struct) for hints.