Prevent file exhaustion from bricking websocket netlayer
This uses wait-for-port in all the areas where reading from the
websocket is occuring. This will timeout the port if nothing is
read after 30 seconds. It also will catch any errors from the accept
call and return #t to the loop instead of propegating the error up and
bricking the netlayer.