Add checks for desc:handoff-receive sessiona and fix certs
The reciever-session used to contain the gifter <-> receiver session
name and the receiver-side was the receiver side in that session. That
is not what it should be or what the spec says, this fixes this by
putting the session information between the receiver and exporter
instead.

Related: https://github.com/ocapn/ocapn-test-suite/issues/12