Fix exceptions in some cases when promise pipelining
In eef1512 it was introduced to use syscaller-<- instead of just sending
a message to get its handling of remote-refrs. This was fixing a problem
where promise pipelining wasn't properly happening as the question
answer machinary was being skipped in this case.

This however introduced a regression where it assumed resolve-me (the
message) resolver was always present. This is of course not the case for
messages sent with <-np and so exceptions would be raised. This fixes
that by checking if resolve-me is a live-refr.