Use new aurie-local-link mactor for local refr in resturation
Aurie uses local-links when resturing local refrs. It initially creates
a promise pair, and then the permanrnt local refr for each near refr It
restores. These local refrs strangely point at promises through the link
mactor created. Once the refr has been rehydrated these are then swapped
out for the real mactor underneed.

While this is already all sorts of weird (local refrs shouldn't point at
promises!), it's also weird at the mactor level too. <mactor:local-link>
is typically used when a promise resolves to an object, not to point at
a promise. It was the best mactor type we had, but now we have a new
aurie-local-link which is a slightly better fit.

This changes the use of local-link to aurie-local-link, making just a
tad more sense (but still very confusing :)).