Fix multiple aurie issues introduced in 0.17
There are several issues introduced:

- Aurie would take a full snapshot of the graph when it wasn't needed
and would not take one when it was needed. This fixes it so that a full
snapshot happens either when we have freshly spawned the vat, or when
we've upgraded the roots.
- make-actormap-read-portrait! had the refrs->val hashmap the wrong way
round (restoring gives you refrs->slots, not slots->refrs).
- Objects were not being properly added to the set of changes Objects
- The logic around if an object is marked as changed was inverted.
- the test in test-vat used #:restore which requires the definer to
tag the restored refr with a version to have upgrading work properly.